Fixes from rebase
